cyberpolygon/requirements.txt
anyio==4.6.0
asgiref==3.8.1
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1
django-allauth==0.61.1
django-allauth-2fa==0.11.1
django-cors-headers==4.4.0
django-filter==24.3
django-otp==1.5.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fido2==1.1.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jsonfield==3.1.0
Markdown==3.5.2
martor==1.6.44
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.1
python-telegram-bot==21.6
python3-openid==3.2.0
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
webencodings==0.5.1
